==> data-raw/util/8-topic_model.Rmd <==
---
jupyter:
  jupytext:
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.2'
      jupytext_version: 1.6.0
  kernelspec:
    display_name: R
    language: R
    name: ir
---

<!-- #region slideshow={"slide_type": "slide"} -->
# Create Topic Model for Incidents
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:13.059805Z', 'iopub.status.busy': '2020-11-10T19:12:13.057266Z', 'iopub.status.idle': '2020-11-10T19:12:15.836580Z', 'shell.execute_reply': '2020-11-10T19:12:15.835969Z'}}
suppressMessages(library(here))
suppressMessages(library(tidyverse))
library(tidytext)
library(SnowballC)
library(topicmodels)
library(ldatuning)
library(here)

assets <- purrr::partial(here, "data-raw", "util", "assets")
data_folder <- purrr::partial(here, "data-raw", ".temp", "data")
```

<!-- #region slideshow={"slide_type": "subslide"} -->
Used data as of Oct 12, 2020.

Original exploration can be found at: https://julianbarg.github.io/spills/inquiries/incident_variance/2020-08-03.slides.html#/
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:15.879844Z', 'iopub.status.busy': '2020-11-10T19:12:15.851223Z', 'iopub.status.idle': '2020-11-10T19:12:16.107848Z', 'shell.execute_reply': '2020-11-10T19:12:16.106587Z'}}
# incidents <- readRDS(data_folder("incidents_merged.rds"))
# readr::write_rds(
#   select(incidents, incident_ID, ID, name, on_offshore, commodity, year,
#          narrative),
#   assets("narratives.rds"))
narratives <- readRDS(assets("narratives.rds"))
head(narratives)
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### We want to grab the company names, so we can remove those from the dataset
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:16.117602Z', 'iopub.status.busy': '2020-11-10T19:12:16.115597Z', 'iopub.status.idle': '2020-11-10T19:12:16.327421Z', 'shell.execute_reply': '2020-11-10T19:12:16.325972Z'}}
company_names <- unique(word(narratives$name, 1))
paste(company_names, collapse = ", ")
```

<!-- #region slideshow={"slide_type": "subslide"} -->
Limite data set to observation period and only retain relevant columns.
<!-- #endregion -->

<!-- #region slideshow={"slide_type": "slide"} -->
## Create document-term matrix
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:16.334978Z', 'iopub.status.busy': '2020-11-10T19:12:16.332335Z', 'iopub.status.idle': '2020-11-10T19:12:16.350303Z', 'shell.execute_reply': '2020-11-10T19:12:16.348815Z'}, slideshow={'slide_type': 'subslide'}}
narratives <- subset(narratives, on_offshore == "onshore" 
                     & commodity %in% c("rpp", "crude") 
                     & year >= 2004 & year < 2020 
                     & !is.na(narrative))
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:16.356363Z', 'iopub.status.busy': '2020-11-10T19:12:16.354658Z', 'iopub.status.idle': '2020-11-10T19:12:21.809657Z', 'shell.execute_reply': '2020-11-10T19:12:21.808767Z'}}
word_counts <- narratives %>%
    unnest_tokens(word, narrative) %>%
    anti_join(data.frame(word = str_to_lower(company_names)), by = "word") %>%
    anti_join(stop_words, by = "word") %>%
    filter(! str_detect(word, "^[0-9]")) %>%
    mutate(word = wordStem(word)) %>%
    count(incident_ID, word, sort = T) %>%
    cast_dtm(document = incident_ID, term = word, 
             value = n)
word_counts
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Run topicmodels
Models based on data as of Oct 12, 2020.
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:21.814173Z', 'iopub.status.busy': '2020-11-10T19:12:21.812853Z', 'iopub.status.idle': '2020-11-10T19:12:21.825099Z', 'shell.execute_reply': '2020-11-10T19:12:21.824149Z'}, slideshow={'slide_type': 'subslide'}}
# results_1 <- FindTopicsNumber(word_counts, 
#                               topics = c(5, 10, 20, 40, 80, 120, 160), 
#                               metrics = c("Griffiths2004", "CaoJuan2009", "Arun2010", "Deveaud2014"),
#                               method = "Gibbs", 
#                               control = list(seed = 532))
# saveRDS(results_1, file = assets("results_1.rds"))
results_1 <- readRDS(assets("results_1.rds"))
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:21.830040Z', 'iopub.status.busy': '2020-11-10T19:12:21.828468Z', 'iopub.status.idle': '2020-11-10T19:12:22.748416Z', 'shell.execute_reply': '2020-11-10T19:12:22.747121Z'}, slideshow={'slide_type': 'subslide'}}
FindTopicsNumber_plot(results_1)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:22.754454Z', 'iopub.status.busy': '2020-11-10T19:12:22.752486Z', 'iopub.status.idle': '2020-11-10T19:12:22.766527Z', 'shell.execute_reply': '2020-11-10T19:12:22.765432Z'}, slideshow={'slide_type': 'subslide'}}
# results_2 <- FindTopicsNumber(word_counts, 
#                             topics = c(15, 18, 19, 20, 21, 22, 25), 
#                             metrics = c("Griffiths2004", "CaoJuan2009", "Arun2010", "Deveaud2014"),
#                             method = "Gibbs", 
#                             control = list(seed = 544))
# saveRDS(results_2, file = assets("results_2.rds"))
results_2 <- readRDS(assets("results_2.rds"))
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:22.771892Z', 'iopub.status.busy': '2020-11-10T19:12:22.770577Z', 'iopub.status.idle': '2020-11-10T19:12:23.293976Z', 'shell.execute_reply': '2020-11-10T19:12:23.292958Z'}, slideshow={'slide_type': 'subslide'}}
FindTopicsNumber_plot(results_2)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:23.299005Z', 'iopub.status.busy': '2020-11-10T19:12:23.296941Z', 'iopub.status.idle': '2020-11-10T19:12:23.311265Z', 'shell.execute_reply': '2020-11-10T19:12:23.310107Z'}, slideshow={'slide_type': 'subslide'}}
# results_3 <- FindTopicsNumber(word_counts, 
#                             topics = seq(15, 35), 
#                             metrics = c("Griffiths2004", "CaoJuan2009", "Arun2010", "Deveaud2014"),
#                             method = "Gibbs", 
#                             control = list(seed = 912))
# saveRDS(results_3, file = assets("results_3.rds"))
results_3 <- readRDS(assets("results_3.rds"))
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:23.315403Z', 'iopub.status.busy': '2020-11-10T19:12:23.314034Z', 'iopub.status.idle': '2020-11-10T19:12:23.842422Z', 'shell.execute_reply': '2020-11-10T19:12:23.840995Z'}, slideshow={'slide_type': 'subslide'}}
FindTopicsNumber_plot(results_3)
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Create model to be used
Models based on data as of Oct 12, 2020.
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:23.847428Z', 'iopub.status.busy': '2020-11-10T19:12:23.846109Z', 'iopub.status.idle': '2020-11-10T19:12:23.912087Z', 'shell.execute_reply': '2020-11-10T19:12:23.910798Z'}}
# narratives_lda <- LDA(word_counts, method = "Gibbs", k = 23, control = list(seed = 912))
# saveRDS(narratives_lda, assets("narratives_lda.rds"))
narratives_lda <- readRDS(assets("narratives_lda.rds"))
narratives_lda
```

<!-- #region slideshow={"slide_type": "subslide"} -->
## Get betas and gammas 
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:23.923631Z', 'iopub.status.busy': '2020-11-10T19:12:23.923083Z', 'iopub.status.idle': '2020-11-10T19:12:23.969522Z', 'shell.execute_reply': '2020-11-10T19:12:23.968408Z'}}
gammas <- tidy(narratives_lda, matrix = "gamma")
head(gammas)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:23.974688Z', 'iopub.status.busy': '2020-11-10T19:12:23.973230Z', 'iopub.status.idle': '2020-11-10T19:12:24.017913Z', 'shell.execute_reply': '2020-11-10T19:12:24.016548Z'}, slideshow={'slide_type': 'subslide'}}
betas <- tidy(narratives_lda, matrix = "beta")
head(betas)
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Save model
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:24.022894Z', 'iopub.status.busy': '2020-11-10T19:12:24.021486Z', 'iopub.status.idle': '2020-11-10T19:12:24.183034Z', 'shell.execute_reply': '2020-11-10T19:12:24.181902Z'}}
saveRDS(betas, data_folder("betas.rds"))
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:24.188121Z', 'iopub.status.busy': '2020-11-10T19:12:24.186769Z', 'iopub.status.idle': '2020-11-10T19:12:24.233551Z', 'shell.execute_reply': '2020-11-10T19:12:24.232389Z'}, slideshow={'slide_type': 'subslide'}}
gammas <- narratives %>%
    mutate(incident_ID = as.character(incident_ID)) %>%
    select(incident_ID, ID, year, commodity, on_offshore) %>%
    right_join(gammas, by = c("incident_ID" = "document"))

head(gammas)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:24.238247Z', 'iopub.status.busy': '2020-11-10T19:12:24.236903Z', 'iopub.status.idle': '2020-11-10T19:12:24.448052Z', 'shell.execute_reply': '2020-11-10T19:12:24.447159Z'}, slideshow={'slide_type': 'subslide'}}
saveRDS(gammas, data_folder("gammas.rds"))
```
