==> data-raw/util/9-interpret_topics.Rmd <==
---
jupyter:
  jupytext:
    formats: ipynb,Rmd
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.2'
      jupytext_version: 1.6.0
  kernelspec:
    display_name: R
    language: R
    name: ir
---

<!-- #region slideshow={"slide_type": "slide"} -->
# Interpret Topics
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:30.562278Z', 'iopub.status.busy': '2020-11-10T19:12:30.560103Z', 'iopub.status.idle': '2020-11-10T19:12:32.943390Z', 'shell.execute_reply': '2020-11-10T19:12:32.942584Z'}, slideshow={'slide_type': 'skip'}}
# library("oildata")
library("incidentmodels")
suppressMessages(library("tidyverse"))
library("tidytext")
suppressMessages(library("magrittr"))
suppressMessages(library("glue"))
suppressMessages(library("here"))

data_folder <- purrr::partial(here, "data-raw", ".temp", "data")
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:32.970562Z', 'iopub.status.busy': '2020-11-10T19:12:32.946207Z', 'iopub.status.idle': '2020-11-10T19:12:33.221017Z', 'shell.execute_reply': '2020-11-10T19:12:33.219739Z'}, slideshow={'slide_type': 'subslide'}}
gammas <- readRDS(data_folder("gammas.rds"))
gammas$label <- as.character(NA)
gammas$cause_topic <- NA
head(gammas)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.226151Z', 'iopub.status.busy': '2020-11-10T19:12:33.224688Z', 'iopub.status.idle': '2020-11-10T19:12:33.291300Z', 'shell.execute_reply': '2020-11-10T19:12:33.289621Z'}, slideshow={'slide_type': 'subslide'}}
head(betas)
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Functions
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.296550Z', 'iopub.status.busy': '2020-11-10T19:12:33.295312Z', 'iopub.status.idle': '2020-11-10T19:12:33.399256Z', 'shell.execute_reply': '2020-11-10T19:12:33.398196Z'}}
top_terms <- betas %>%
    arrange(desc(beta)) %>%
    group_by(topic) %>%
    slice_head(n = 8) %>%
    ungroup() %>%
    arrange(topic, -beta)

head(top_terms)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.404393Z', 'iopub.status.busy': '2020-11-10T19:12:33.402731Z', 'iopub.status.idle': '2020-11-10T19:12:33.531147Z', 'shell.execute_reply': '2020-11-10T19:12:33.530238Z'}}
get_terms <- function(topic) {
    filter(top_terms, topic == {{ topic }})
}
narratives <- select(incidents, incident_ID, narrative) %>%
    mutate(incident_ID = as.character(incident_ID))
get_narrative <- function(topic) {
    gammas %>%
        filter(topic == {{ topic }}) %>%
        filter(gamma == max(gamma)) %>%
        left_join(narratives, by = "incident_ID") %$%
        glue("Incident ID: {incident_ID}
              Operator ID: {ID}

              {narrative}")
}
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Topics 1-5
<!-- #endregion -->

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 1
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.536341Z', 'iopub.status.busy': '2020-11-10T19:12:33.534984Z', 'iopub.status.idle': '2020-11-10T19:12:33.558773Z', 'shell.execute_reply': '2020-11-10T19:12:33.557524Z'}}
get_terms(topic = 1)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.562500Z', 'iopub.status.busy': '2020-11-10T19:12:33.561524Z', 'iopub.status.idle': '2020-11-10T19:12:33.595095Z', 'shell.execute_reply': '2020-11-10T19:12:33.593534Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(1))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic seems to be related to the management of incident reports**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.602413Z', 'iopub.status.busy': '2020-11-10T19:12:33.600511Z', 'iopub.status.idle': '2020-11-10T19:12:33.771532Z', 'shell.execute_reply': '2020-11-10T19:12:33.769893Z'}}
gammas[gammas$topic == 1, ]$label <- "report_mngt"
gammas[gammas$topic == 1, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 2
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.777204Z', 'iopub.status.busy': '2020-11-10T19:12:33.775423Z', 'iopub.status.idle': '2020-11-10T19:12:33.805672Z', 'shell.execute_reply': '2020-11-10T19:12:33.804193Z'}}
get_terms(topic = 2)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.812033Z', 'iopub.status.busy': '2020-11-10T19:12:33.810644Z', 'iopub.status.idle': '2020-11-10T19:12:33.851657Z', 'shell.execute_reply': '2020-11-10T19:12:33.850117Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(2))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to the management of spills**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.857590Z', 'iopub.status.busy': '2020-11-10T19:12:33.855860Z', 'iopub.status.idle': '2020-11-10T19:12:33.883255Z', 'shell.execute_reply': '2020-11-10T19:12:33.882169Z'}}
gammas[gammas$topic == 2, ]$label <- "spill_mngt"
gammas[gammas$topic == 2, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 3
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.889043Z', 'iopub.status.busy': '2020-11-10T19:12:33.887227Z', 'iopub.status.idle': '2020-11-10T19:12:33.923393Z', 'shell.execute_reply': '2020-11-10T19:12:33.921993Z'}}
get_terms(topic = 3)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.928366Z', 'iopub.status.busy': '2020-11-10T19:12:33.927025Z', 'iopub.status.idle': '2020-11-10T19:12:33.953101Z', 'shell.execute_reply': '2020-11-10T19:12:33.951772Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(3))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to contractors and excavation**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.957896Z', 'iopub.status.busy': '2020-11-10T19:12:33.956627Z', 'iopub.status.idle': '2020-11-10T19:12:33.977583Z', 'shell.execute_reply': '2020-11-10T19:12:33.976529Z'}}
gammas[gammas$topic == 3, ]$label <- "excavation"
gammas[gammas$topic == 3, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 4
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:33.982367Z', 'iopub.status.busy': '2020-11-10T19:12:33.981053Z', 'iopub.status.idle': '2020-11-10T19:12:34.004681Z', 'shell.execute_reply': '2020-11-10T19:12:34.003460Z'}}
get_terms(topic = 4)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.009517Z', 'iopub.status.busy': '2020-11-10T19:12:34.008261Z', 'iopub.status.idle': '2020-11-10T19:12:34.040589Z', 'shell.execute_reply': '2020-11-10T19:12:34.039445Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(4))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to storms, water and related damages**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.045339Z', 'iopub.status.busy': '2020-11-10T19:12:34.044068Z', 'iopub.status.idle': '2020-11-10T19:12:34.197214Z', 'shell.execute_reply': '2020-11-10T19:12:34.196168Z'}}
gammas[gammas$topic == 4, ]$label <- "water"
gammas[gammas$topic == 4, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 5
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.201958Z', 'iopub.status.busy': '2020-11-10T19:12:34.200712Z', 'iopub.status.idle': '2020-11-10T19:12:34.223654Z', 'shell.execute_reply': '2020-11-10T19:12:34.222471Z'}}
get_terms(topic = 5)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.228355Z', 'iopub.status.busy': '2020-11-10T19:12:34.227072Z', 'iopub.status.idle': '2020-11-10T19:12:34.254852Z', 'shell.execute_reply': '2020-11-10T19:12:34.253567Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(5))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to remote operation of pipelines**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.259604Z', 'iopub.status.busy': '2020-11-10T19:12:34.258304Z', 'iopub.status.idle': '2020-11-10T19:12:34.279149Z', 'shell.execute_reply': '2020-11-10T19:12:34.278081Z'}}
gammas[gammas$topic == 5, ]$label <- "control_center"
gammas[gammas$topic == 5, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Topics 6-10
<!-- #endregion -->

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 6
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.284014Z', 'iopub.status.busy': '2020-11-10T19:12:34.282724Z', 'iopub.status.idle': '2020-11-10T19:12:34.305968Z', 'shell.execute_reply': '2020-11-10T19:12:34.304790Z'}}
get_terms(topic = 6)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.310711Z', 'iopub.status.busy': '2020-11-10T19:12:34.309393Z', 'iopub.status.idle': '2020-11-10T19:12:34.335427Z', 'shell.execute_reply': '2020-11-10T19:12:34.334160Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(6))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to service and repair**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.340330Z', 'iopub.status.busy': '2020-11-10T19:12:34.339000Z', 'iopub.status.idle': '2020-11-10T19:12:34.368990Z', 'shell.execute_reply': '2020-11-10T19:12:34.367268Z'}}
gammas[gammas$topic == 6, ]$label <- "service"
gammas[gammas$topic == 6, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 7
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.375099Z', 'iopub.status.busy': '2020-11-10T19:12:34.373379Z', 'iopub.status.idle': '2020-11-10T19:12:34.403292Z', 'shell.execute_reply': '2020-11-10T19:12:34.401462Z'}}
get_terms(topic = 7)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.408268Z', 'iopub.status.busy': '2020-11-10T19:12:34.406749Z', 'iopub.status.idle': '2020-11-10T19:12:34.432602Z', 'shell.execute_reply': '2020-11-10T19:12:34.431287Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(7))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to pumps and their components**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.437781Z', 'iopub.status.busy': '2020-11-10T19:12:34.436235Z', 'iopub.status.idle': '2020-11-10T19:12:34.462249Z', 'shell.execute_reply': '2020-11-10T19:12:34.461328Z'}}
gammas[gammas$topic == 7, ]$label <- "pumps"
gammas[gammas$topic == 7, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 8
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.466829Z', 'iopub.status.busy': '2020-11-10T19:12:34.465510Z', 'iopub.status.idle': '2020-11-10T19:12:34.489166Z', 'shell.execute_reply': '2020-11-10T19:12:34.487995Z'}}
get_terms(topic = 8)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.492925Z', 'iopub.status.busy': '2020-11-10T19:12:34.491954Z', 'iopub.status.idle': '2020-11-10T19:12:34.512871Z', 'shell.execute_reply': '2020-11-10T19:12:34.511764Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(8))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to procedures**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.517767Z', 'iopub.status.busy': '2020-11-10T19:12:34.516242Z', 'iopub.status.idle': '2020-11-10T19:12:34.537913Z', 'shell.execute_reply': '2020-11-10T19:12:34.536771Z'}}
gammas[gammas$topic == 8, ]$label <- "procedures"
gammas[gammas$topic == 8, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 9
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.542026Z', 'iopub.status.busy': '2020-11-10T19:12:34.540959Z', 'iopub.status.idle': '2020-11-10T19:12:34.558846Z', 'shell.execute_reply': '2020-11-10T19:12:34.557973Z'}}
get_terms(topic = 9)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.563129Z', 'iopub.status.busy': '2020-11-10T19:12:34.561801Z', 'iopub.status.idle': '2020-11-10T19:12:34.588129Z', 'shell.execute_reply': '2020-11-10T19:12:34.587082Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(9))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to the commodity transported**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.593075Z', 'iopub.status.busy': '2020-11-10T19:12:34.591411Z', 'iopub.status.idle': '2020-11-10T19:12:34.623113Z', 'shell.execute_reply': '2020-11-10T19:12:34.622042Z'}}
gammas[gammas$topic == 9, ]$label <- "commodity"
gammas[gammas$topic == 9, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 10
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.626988Z', 'iopub.status.busy': '2020-11-10T19:12:34.626084Z', 'iopub.status.idle': '2020-11-10T19:12:34.644010Z', 'shell.execute_reply': '2020-11-10T19:12:34.643190Z'}}
get_terms(topic = 10)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.648173Z', 'iopub.status.busy': '2020-11-10T19:12:34.646909Z', 'iopub.status.idle': '2020-11-10T19:12:34.666886Z', 'shell.execute_reply': '2020-11-10T19:12:34.665868Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(10))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to testing and monitoring**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.671983Z', 'iopub.status.busy': '2020-11-10T19:12:34.670579Z', 'iopub.status.idle': '2020-11-10T19:12:34.692099Z', 'shell.execute_reply': '2020-11-10T19:12:34.690957Z'}}
gammas[gammas$topic == 10, ]$label <- "monitoring"
gammas[gammas$topic == 10, ]$cause_topic <- T
```

## Topics 11-15

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 11
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.696670Z', 'iopub.status.busy': '2020-11-10T19:12:34.695336Z', 'iopub.status.idle': '2020-11-10T19:12:34.713354Z', 'shell.execute_reply': '2020-11-10T19:12:34.712500Z'}}
get_terms(topic = 11)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.717572Z', 'iopub.status.busy': '2020-11-10T19:12:34.716289Z', 'iopub.status.idle': '2020-11-10T19:12:34.736317Z', 'shell.execute_reply': '2020-11-10T19:12:34.735259Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(11))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related flow, pressure and relief**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.740960Z', 'iopub.status.busy': '2020-11-10T19:12:34.739630Z', 'iopub.status.idle': '2020-11-10T19:12:34.885332Z', 'shell.execute_reply': '2020-11-10T19:12:34.883862Z'}}
gammas[gammas$topic == 11, ]$label <- "pressure"
gammas[gammas$topic == 11, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 12
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.889540Z', 'iopub.status.busy': '2020-11-10T19:12:34.888367Z', 'iopub.status.idle': '2020-11-10T19:12:34.910090Z', 'shell.execute_reply': '2020-11-10T19:12:34.908819Z'}}
get_terms(topic = 12)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.915691Z', 'iopub.status.busy': '2020-11-10T19:12:34.913786Z', 'iopub.status.idle': '2020-11-10T19:12:34.937162Z', 'shell.execute_reply': '2020-11-10T19:12:34.936157Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(12))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to corrosion of pipes**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.940669Z', 'iopub.status.busy': '2020-11-10T19:12:34.939868Z', 'iopub.status.idle': '2020-11-10T19:12:34.954669Z', 'shell.execute_reply': '2020-11-10T19:12:34.954006Z'}}
gammas[gammas$topic == 12, ]$label <- "corrosion"
gammas[gammas$topic == 12, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 13
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.958705Z', 'iopub.status.busy': '2020-11-10T19:12:34.957502Z', 'iopub.status.idle': '2020-11-10T19:12:34.975191Z', 'shell.execute_reply': '2020-11-10T19:12:34.974191Z'}}
get_terms(topic = 13)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:34.979428Z', 'iopub.status.busy': '2020-11-10T19:12:34.978161Z', 'iopub.status.idle': '2020-11-10T19:12:35.002015Z', 'shell.execute_reply': '2020-11-10T19:12:35.000889Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(13))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to gaskets and related components**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.005623Z', 'iopub.status.busy': '2020-11-10T19:12:35.004777Z', 'iopub.status.idle': '2020-11-10T19:12:35.026322Z', 'shell.execute_reply': '2020-11-10T19:12:35.025629Z'}}
gammas[gammas$topic == 13, ]$label <- "gaskets"
gammas[gammas$topic == 13, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 14
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.030593Z', 'iopub.status.busy': '2020-11-10T19:12:35.029397Z', 'iopub.status.idle': '2020-11-10T19:12:35.047283Z', 'shell.execute_reply': '2020-11-10T19:12:35.046318Z'}}
get_terms(topic = 14)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.051635Z', 'iopub.status.busy': '2020-11-10T19:12:35.050235Z', 'iopub.status.idle': '2020-11-10T19:12:35.074263Z', 'shell.execute_reply': '2020-11-10T19:12:35.073012Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(14))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to reporting of the spill**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.077898Z', 'iopub.status.busy': '2020-11-10T19:12:35.077083Z', 'iopub.status.idle': '2020-11-10T19:12:35.092865Z', 'shell.execute_reply': '2020-11-10T19:12:35.092227Z'}}
gammas[gammas$topic == 14, ]$label <- "report"
gammas[gammas$topic == 14, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 15
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.097082Z', 'iopub.status.busy': '2020-11-10T19:12:35.095874Z', 'iopub.status.idle': '2020-11-10T19:12:35.113311Z', 'shell.execute_reply': '2020-11-10T19:12:35.112341Z'}}
get_terms(topic = 15)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.117485Z', 'iopub.status.busy': '2020-11-10T19:12:35.116271Z', 'iopub.status.idle': '2020-11-10T19:12:35.139220Z', 'shell.execute_reply': '2020-11-10T19:12:35.138089Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(15))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to the personnels response on site**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.143579Z', 'iopub.status.busy': '2020-11-10T19:12:35.142322Z', 'iopub.status.idle': '2020-11-10T19:12:35.162090Z', 'shell.execute_reply': '2020-11-10T19:12:35.161067Z'}}
gammas[gammas$topic == 15, ]$label <- "response"
gammas[gammas$topic == 15, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Topics 16-20
<!-- #endregion -->

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 16
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.166294Z', 'iopub.status.busy': '2020-11-10T19:12:35.165440Z', 'iopub.status.idle': '2020-11-10T19:12:35.182254Z', 'shell.execute_reply': '2020-11-10T19:12:35.181391Z'}}
get_terms(topic = 16)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.186395Z', 'iopub.status.busy': '2020-11-10T19:12:35.185467Z', 'iopub.status.idle': '2020-11-10T19:12:35.209004Z', 'shell.execute_reply': '2020-11-10T19:12:35.208152Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(16))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to valves and related equipment**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.213040Z', 'iopub.status.busy': '2020-11-10T19:12:35.211843Z', 'iopub.status.idle': '2020-11-10T19:12:35.227720Z', 'shell.execute_reply': '2020-11-10T19:12:35.226830Z'}}
gammas[gammas$topic == 16, ]$label <- "valve"
gammas[gammas$topic == 16, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 17
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.233101Z', 'iopub.status.busy': '2020-11-10T19:12:35.232567Z', 'iopub.status.idle': '2020-11-10T19:12:35.254473Z', 'shell.execute_reply': '2020-11-10T19:12:35.253343Z'}}
get_terms(topic = 17)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.258033Z', 'iopub.status.busy': '2020-11-10T19:12:35.257141Z', 'iopub.status.idle': '2020-11-10T19:12:35.276162Z', 'shell.execute_reply': '2020-11-10T19:12:35.275053Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(17))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to the leak and clean up**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.280308Z', 'iopub.status.busy': '2020-11-10T19:12:35.279076Z', 'iopub.status.idle': '2020-11-10T19:12:35.302153Z', 'shell.execute_reply': '2020-11-10T19:12:35.301228Z'}}
gammas[gammas$topic == 17, ]$label <- "leak"
gammas[gammas$topic == 17, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 18
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.306432Z', 'iopub.status.busy': '2020-11-10T19:12:35.305190Z', 'iopub.status.idle': '2020-11-10T19:12:35.327018Z', 'shell.execute_reply': '2020-11-10T19:12:35.325786Z'}}
get_terms(topic = 18)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.331644Z', 'iopub.status.busy': '2020-11-10T19:12:35.330432Z', 'iopub.status.idle': '2020-11-10T19:12:35.354184Z', 'shell.execute_reply': '2020-11-10T19:12:35.353069Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(18))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to soil, contamination and cleanup**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.357947Z', 'iopub.status.busy': '2020-11-10T19:12:35.357026Z', 'iopub.status.idle': '2020-11-10T19:12:35.372315Z', 'shell.execute_reply': '2020-11-10T19:12:35.371669Z'}}
gammas[gammas$topic == 18, ]$label <- "contamination"
gammas[gammas$topic == 18, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 19
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.376385Z', 'iopub.status.busy': '2020-11-10T19:12:35.375217Z', 'iopub.status.idle': '2020-11-10T19:12:35.394855Z', 'shell.execute_reply': '2020-11-10T19:12:35.393797Z'}}
get_terms(topic = 19)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.399078Z', 'iopub.status.busy': '2020-11-10T19:12:35.397849Z', 'iopub.status.idle': '2020-11-10T19:12:35.421058Z', 'shell.execute_reply': '2020-11-10T19:12:35.419962Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(19))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to the release of oil**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.425233Z', 'iopub.status.busy': '2020-11-10T19:12:35.423981Z', 'iopub.status.idle': '2020-11-10T19:12:35.439733Z', 'shell.execute_reply': '2020-11-10T19:12:35.439077Z'}}
gammas[gammas$topic == 19, ]$label <- "release"
gammas[gammas$topic == 19, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 20
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.445009Z', 'iopub.status.busy': '2020-11-10T19:12:35.443046Z', 'iopub.status.idle': '2020-11-10T19:12:35.462671Z', 'shell.execute_reply': '2020-11-10T19:12:35.461597Z'}}
get_terms(topic = 20)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.467026Z', 'iopub.status.busy': '2020-11-10T19:12:35.465727Z', 'iopub.status.idle': '2020-11-10T19:12:35.497359Z', 'shell.execute_reply': '2020-11-10T19:12:35.496329Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(20))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic contains words to describe a spill**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.501724Z', 'iopub.status.busy': '2020-11-10T19:12:35.500284Z', 'iopub.status.idle': '2020-11-10T19:12:35.515837Z', 'shell.execute_reply': '2020-11-10T19:12:35.515189Z'}}
gammas[gammas$topic == 20, ]$label <- "spill"
gammas[gammas$topic == 20, ]$cause_topic <- F
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Topics 21-23
<!-- #endregion -->

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 21
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.520148Z', 'iopub.status.busy': '2020-11-10T19:12:35.518938Z', 'iopub.status.idle': '2020-11-10T19:12:35.536323Z', 'shell.execute_reply': '2020-11-10T19:12:35.535337Z'}}
get_terms(topic = 21)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.540721Z', 'iopub.status.busy': '2020-11-10T19:12:35.539437Z', 'iopub.status.idle': '2020-11-10T19:12:35.563570Z', 'shell.execute_reply': '2020-11-10T19:12:35.562395Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(21))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to fire and emergencies**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.568738Z', 'iopub.status.busy': '2020-11-10T19:12:35.567225Z', 'iopub.status.idle': '2020-11-10T19:12:35.598384Z', 'shell.execute_reply': '2020-11-10T19:12:35.597624Z'}}
gammas[gammas$topic == 21, ]$label <- "fire"
gammas[gammas$topic == 21, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 22
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.603152Z', 'iopub.status.busy': '2020-11-10T19:12:35.602219Z', 'iopub.status.idle': '2020-11-10T19:12:35.626093Z', 'shell.execute_reply': '2020-11-10T19:12:35.625013Z'}}
get_terms(topic = 22)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.629577Z', 'iopub.status.busy': '2020-11-10T19:12:35.628746Z', 'iopub.status.idle': '2020-11-10T19:12:35.646290Z', 'shell.execute_reply': '2020-11-10T19:12:35.645518Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(22))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to tanks and related facilities and equipment**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.650464Z', 'iopub.status.busy': '2020-11-10T19:12:35.649190Z', 'iopub.status.idle': '2020-11-10T19:12:35.665063Z', 'shell.execute_reply': '2020-11-10T19:12:35.664220Z'}}
gammas[gammas$topic == 22, ]$label <- "tanks"
gammas[gammas$topic == 22, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "subslide"} -->
### Topic 23
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.669353Z', 'iopub.status.busy': '2020-11-10T19:12:35.668221Z', 'iopub.status.idle': '2020-11-10T19:12:35.690170Z', 'shell.execute_reply': '2020-11-10T19:12:35.689119Z'}}
get_terms(topic = 23)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.694587Z', 'iopub.status.busy': '2020-11-10T19:12:35.693428Z', 'iopub.status.idle': '2020-11-10T19:12:35.718266Z', 'shell.execute_reply': '2020-11-10T19:12:35.717068Z'}, slideshow={'slide_type': 'subslide'}}
print(get_narrative(23))
```

<!-- #region slideshow={"slide_type": "subslide"} -->
**This topic is related to cracks and other kinds of failures**
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.723364Z', 'iopub.status.busy': '2020-11-10T19:12:35.721879Z', 'iopub.status.idle': '2020-11-10T19:12:35.742070Z', 'shell.execute_reply': '2020-11-10T19:12:35.741116Z'}}
gammas[gammas$topic == 23, ]$label <- "crack"
gammas[gammas$topic == 23, ]$cause_topic <- T
```

<!-- #region slideshow={"slide_type": "slide"} -->
## Save results
<!-- #endregion -->

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.746790Z', 'iopub.status.busy': '2020-11-10T19:12:35.745456Z', 'iopub.status.idle': '2020-11-10T19:12:35.770473Z', 'shell.execute_reply': '2020-11-10T19:12:35.769019Z'}}
head(gammas)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.775166Z', 'iopub.status.busy': '2020-11-10T19:12:35.773839Z', 'iopub.status.idle': '2020-11-10T19:12:35.834715Z', 'shell.execute_reply': '2020-11-10T19:12:35.833531Z'}, slideshow={'slide_type': 'subslide'}}
table(unique(paste(gammas$topic, gammas$label)))
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.839468Z', 'iopub.status.busy': '2020-11-10T19:12:35.838181Z', 'iopub.status.idle': '2020-11-10T19:12:35.873301Z', 'shell.execute_reply': '2020-11-10T19:12:35.872058Z'}, slideshow={'slide_type': 'subslide'}}
labels <- distinct(select(gammas, topic, label, cause_topic))
labels$topic <- as.character(labels$topic)
head(labels)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.878216Z', 'iopub.status.busy': '2020-11-10T19:12:35.876997Z', 'iopub.status.idle': '2020-11-10T19:12:35.904129Z', 'shell.execute_reply': '2020-11-10T19:12:35.902705Z'}}
head(gammas)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.909639Z', 'iopub.status.busy': '2020-11-10T19:12:35.908379Z', 'iopub.status.idle': '2020-11-10T19:12:35.921624Z', 'shell.execute_reply': '2020-11-10T19:12:35.920577Z'}}
write_rds(labels, data_folder("labels.rds"))
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:35.926794Z', 'iopub.status.busy': '2020-11-10T19:12:35.925337Z', 'iopub.status.idle': '2020-11-10T19:12:36.091948Z', 'shell.execute_reply': '2020-11-10T19:12:36.091096Z'}}
incidents <- readRDS(data_folder("incidents_merged.rds"))
incidents_gammas <- gammas %>%
    select(incident_ID, topic, gamma) %>%
    pivot_wider(incident_ID, names_from = topic, values_from = gamma, names_prefix = "topic_") %>%
    right_join(incidents) %>%
    select(-starts_with("topic_"), starts_with("topic_")) # Move topic columns to back
head(incidents_gammas)
```

```{r execution={'iopub.execute_input': '2020-11-10T19:12:36.096761Z', 'iopub.status.busy': '2020-11-10T19:12:36.095452Z', 'iopub.status.idle': '2020-11-10T19:12:36.146308Z', 'shell.execute_reply': '2020-11-10T19:12:36.145085Z'}, slideshow={'slide_type': 'subslide'}}
write_rds(incidents_gammas, data_folder("incidents_topics.rds"))
```
